requirements.txt
python-telegram-bot~=22.3
python-dotenv~=1.1.1